Rename program requisition to program requisition settings
